pyproject.toml
[tool.poetry]
name = "scrapes-lib"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "scrapes_lib"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
parsel = "^1.7.0"
requests = ">2.25,"
uplink = "^0.9.7"
pendulum = "^2.1.2"
sqlalchemy = "^1.4.44"
mysqlclient = "^2.1.1"
pymysql = "^1.0.2"
cryptography = "^38.0.3"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.6.0"
pyyaml = "^6.0"
tinybird-cli = "^1.0.0b210"


[tool.poetry.group.dev.dependencies]
ipython = "^8.6.0"
black = "^22.10.0"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
